Renamed packages to follow naming conventions
