+++ src/main/java/entities/Gradable.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 public interface Gradable {
     /**
