+++ src/main/java/entities/InstructorUser.java
@@ -1,7 +1,6 @@
-package Entities;
+package entities;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public class InstructorUser extends User {
