+++ src/main/java/entities/StudentUser.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import java.time.LocalDateTime;
 import java.util.*;
