+++ src/main/java/entities/Task.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import java.io.Serializable;
