+++ src/main/java/entities/Timeblockable.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 import java.time.LocalDateTime;
 public interface Timeblockable {
     /**
