+++ src/main/java/entities/UserWhisperer.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 // Entity
