+++ src/main/java/entities/internalUserList.java
@@ -1,4 +1,4 @@
-package Entities;
+package entities;
 
 import java.util.ArrayList;
