+++ src/main/java/read_write/ReadWriter.java
@@ -1,4 +1,4 @@
-package ReadWrite;
+package read_write;
 
 import java.io.IOException;
